[csa] Define CallFunctionTemplate builtins' argc parameter as Int32
This unifies calling conventions with other similar builtins.

Bug: chromium:1445925
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4637083
Auto-Submit: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Nico Hartmann <nicohartmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#88468}

+++ src/codegen/interface-descriptors.h
@@ -1420,7 +1420,7 @@ class CallFunctionTemplateDescriptor
  public:
   DEFINE_PARAMETERS_VARARGS(kFunctionTemplateInfo, kArgumentsCount)
   DEFINE_PARAMETER_TYPES(MachineType::AnyTagged(),  // kFunctionTemplateInfo
-                         MachineType::IntPtr())     // kArgumentsCount
+                         MachineType::Int32())      // kArgumentsCount
   DECLARE_DESCRIPTOR(CallFunctionTemplateDescriptor)
 
   static constexpr inline auto registers();
